// model/accountData.ts
/**
 * PostPuma - OpenAPI 3.0
 * PostPuma API specifications
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@postpuma.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class AccountData {
    'unionId'?: string;
    'isPrivate'?: boolean;
    'duetDisabled'?: boolean;
    'privacyLevels'?: Array<string>;
    'stitchDisabled'?: boolean;
    'commentDisabled'?: boolean;
    'maxVideoPostDurationSec'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "unionId",
            "baseName": "union_id",
            "type": "string"
        },
        {
            "name": "isPrivate",
            "baseName": "is_private",
            "type": "boolean"
        },
        {
            "name": "duetDisabled",
            "baseName": "duet_disabled",
            "type": "boolean"
        },
        {
            "name": "privacyLevels",
            "baseName": "privacy_levels",
            "type": "Array<string>"
        },
        {
            "name": "stitchDisabled",
            "baseName": "stitch_disabled",
            "type": "boolean"
        },
        {
            "name": "commentDisabled",
            "baseName": "comment_disabled",
            "type": "boolean"
        },
        {
            "name": "maxVideoPostDurationSec",
            "baseName": "max_video_post_duration_sec",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return AccountData.attributeTypeMap;
    }
}