// model/createPostRequest.ts
/**
 * PostPuma - OpenAPI 3.0
 * PostPuma API specifications
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@postpuma.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Version } from './version';

export class CreatePostRequest {
    'date'?: string;
    'time'?: string;
    'timezone'?: string;
    'schedule'?: boolean;
    'scheduleNow'?: boolean;
    'queue'?: boolean;
    'accounts'?: Array<number>;
    'tags'?: Array<number>;
    'versions'?: Array<Version>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "date",
            "baseName": "date",
            "type": "string"
        },
        {
            "name": "time",
            "baseName": "time",
            "type": "string"
        },
        {
            "name": "timezone",
            "baseName": "timezone",
            "type": "string"
        },
        {
            "name": "schedule",
            "baseName": "schedule",
            "type": "boolean"
        },
        {
            "name": "scheduleNow",
            "baseName": "schedule_now",
            "type": "boolean"
        },
        {
            "name": "queue",
            "baseName": "queue",
            "type": "boolean"
        },
        {
            "name": "accounts",
            "baseName": "accounts",
            "type": "Array<number>"
        },
        {
            "name": "tags",
            "baseName": "tags",
            "type": "Array<number>"
        },
        {
            "name": "versions",
            "baseName": "versions",
            "type": "Array<Version>"
        }    ];

    static getAttributeTypeMap() {
        return CreatePostRequest.attributeTypeMap;
    }
}