// model/deletePostRequest.ts
/**
 * PostPuma - OpenAPI 3.0
 * PostPuma API specifications
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@postpuma.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class DeletePostRequest {
    /**
    * Whether to move items to trash or not.
    */
    'trash'?: boolean;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "trash",
            "baseName": "trash",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return DeletePostRequest.attributeTypeMap;
    }
}