// model/listMediaFiles200Response.ts
/**
 * PostPuma - OpenAPI 3.0
 * PostPuma API specifications
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@postpuma.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { MediaFile } from './mediaFile';

export class ListMediaFiles200Response {
    'data'?: Array<MediaFile>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "data",
            "baseName": "data",
            "type": "Array<MediaFile>"
        }    ];

    static getAttributeTypeMap() {
        return ListMediaFiles200Response.attributeTypeMap;
    }
}